[PR87874] avoid const-wide-int subreg in LRA
Just like CONST_INT, CONST_WIDE_INT is VOIDmode, so LRA might be
tempted to build a SUBREG to "convert" it to the wanted mode.  That's
no use.  Test for CONST_SCALAR_INT_P instead of CONST_INT_P so that we
skip the subreg creation for both.

for  gcc/ChangeLog

	PR rtl-optimization/87874
	* lra.c (lra_substitute_pseudo): Do not create a subreg for
	const wide ints.

for  gcc/testsuite/ChangeLog

	PR rtl-optimization/87874
	* gcc.dg/pr87874.c: New.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@265860 138bc75d-0d04-0410-961f-82ee72b054a4