combine: More make_more_copies
This makes make_more_copies do what its documentation says, that is,
only make an intermediate pseudo if copying to a pseudo.

This regressed generated code quality when we didn't keep the original
notes that were on the copy, but since r265582 we do, and only allowing
pseudos now is a win.  It also simplifies the code.


	* combine.c (make_more_copies): Only make an intermediate copy if the
	dest of a move is a pseudo.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@266004 138bc75d-0d04-0410-961f-82ee72b054a4