[Ada] Remove couple of recently added dead tests
N_Quantified_Expression and N_Iterated_Component_Association are
unrelated nodes that cannot appear in the same context: the former can
appear wherever an expression node is acceptable whereas the latter can
appear only as an element of a component association list.  So a test
combining both most likely contains a dead arm and this change removes a
couple of them.

No functional changes.

2018-11-14  Eric Botcazou  <ebotcazou@adacore.com>

gcc/ada/

	* exp_aggr.adb (Check_Static_Components): Remove dead test.
	(Flatten): Likewise.  Move comment around.

git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@266133 138bc75d-0d04-0410-961f-82ee72b054a4