When running the testsuite on boards that can't report an error status DejaGNU uses a special wrapper to print the exit code on stdout and parses stdout to find whether an execution failed or passed.
In testcases that use "freopen (..., ..., stdout)" this special line is
printed to the alternate location described in the freopen call and
DejaGNU can't find the error code.

This results in DejaGNU using a default return status of 2 and the test
failing.

This patch skips the two testcases that use freopen on stdout when
testing a board that requires this wrapper.


Testing done by running these two tests on arm-none-eabi cross build and
observing that they are unsupported.

gcc/testsuite/ChangeLog:

2018-11-16  Matthew Malcomson  <matthew.malcomson@arm.com>

	* gcc.c-torture/execute/printf-2.c: Skip on wrapped boards.
	* gcc.c-torture/execute/user-printf.c: Likewise.


git-svn-id: svn://gcc.gnu.org/svn/gcc/trunk@266214 138bc75d-0d04-0410-961f-82ee72b054a4